refactor(activerecord): rename test-setup-ar.ts to cases/helper.ts
Mirrors Rails' activerecord/test/cases/helper.rb, which this file is the
trails counterpart of. Single mechanical rename: no behavior change, same
settings in the same order, still the second setupFile (after
test-setup-worker-db.ts, before test-setup-dy.ts).

Updates the vitest setupFiles entry and boot-order prose, the eslint
no-raw-sql scope-out glob and its hardcoded /test-setup-*.ts/ regex (a
cases/helper.ts path no longer matches), the file's own stale header
referring to a sibling test-setup.ts that no longer exists, and the
prose references across the package.
