+++ eslint/no-raw-sql.mjs
@@ -41,6 +41,8 @@ function isExcludedPath(rel) {
   if (/(^|\/)test-helpers\//.test(rel)) return true;
   if (/(^|\/)support\//.test(rel)) return true;
   if (/(^|\/)test-setup-[^/]*\.ts$/.test(rel)) return true;
+  if (/(^|\/)cases\/helper\.ts$/.test(rel)) return true;
+  if (/(^|\/)cases\/helper\.ts$/.test(rel)) return true;
   return false;
 }
